# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "additionalProperties": false,
        "description": "[`Account`](/architecture/interfaces.md#account), the response of `API-21`, `API-23` and\n`API-24`.",
        "properties": {
          "active_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run Id"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "attribute_origin": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attribute Origin",
            "type": "object"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "crunchbase_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crunchbase Id"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "employee_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Count"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "last_refreshed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refreshed At"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_refresh_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Refresh At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "operational_complexity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountOperationalComplexity"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "$ref": "#/components/schemas/AccountOrigin"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountParent"
              },
              {
                "type": "null"
              }
            ]
          },
          "revenue_eur": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Eur"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AccountSourceItem"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          }
        },
        "required": [
          "id",
          "name",
          "domain",
          "country_code",
          "industry",
          "status",
          "origin",
          "last_refreshed_at",
          "active_run_id",
          "employee_count",
          "revenue_eur",
          "operational_complexity",
          "attribute_origin",
          "parent",
          "crunchbase_id",
          "linkedin_url",
          "notes",
          "aliases",
          "sources",
          "next_refresh_at"
        ],
        "title": "Account",
        "type": "object"
      },
...
      "AccountOperationalComplexity": {
        "description": "`account.operational_complexity`.",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "AccountOperationalComplexity",
        "type": "string"
      },
      "AccountOrigin": {
        "description": "`account.origin`.",
        "enum": [
          "IMPORTED",
          "MANUAL",
          "DISCOVERED"
        ],
        "title": "AccountOrigin",
        "type": "string"
      },
      "AccountParent": {
        "additionalProperties": false,
        "description": "`Account.parent`.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "AccountParent",
        "type": "object"
      },
...
      "AccountSourceItem": {
        "additionalProperties": false,
        "description": "One entry of `Account.sources`.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AccountSourceKind"
          },
          "origin": {
            "$ref": "#/components/schemas/AccountSourceOrigin"
          },
          "status": {
            "$ref": "#/components/schemas/AccountSourceStatus"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "kind",
          "origin",
          "status"
        ],
        "title": "AccountSourceItem",
        "type": "object"
      },
      "AccountSourceKind": {
        "description": "`account_source.kind`.",
        "enum": [
          "WEBSITE",
          "NEWSROOM",
          "INVESTOR_RELATIONS",
          "CAREERS",
          "RSS_FEED"
        ],
        "title": "AccountSourceKind",
        "type": "string"
      },
      "AccountSourceOrigin": {
        "description": "`account_source.origin`.",
        "enum": [
          "MANUAL",
          "DETECTED"
        ],
        "title": "AccountSourceOrigin",
        "type": "string"
      },
      "AccountSourceStatus": {
        "description": "`account_source.status`.",
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ],
        "title": "AccountSourceStatus",
        "type": "string"
      },
      "AccountStatus": {
        "description": "`account.status`.",
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ],
        "title": "AccountStatus",
        "type": "string"
      },
...
      "CandidateDecision": {
        "additionalProperties": false,
        "description": "[`CandidateDecision`](/architecture/interfaces.md#candidatedecision).",
        "properties": {
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "CandidateDecision",
        "type": "object"
      },
...
      "Dependency": {
        "description": "`details.dependency` of `UPSTREAM_UNAVAILABLE`, exactly the\n[Dependencies](/architecture/interfaces.md#conventions) table's column: a wire-only enum,\nowned by interfaces and defined here once.",
        "enum": [
          "DATABASE",
          "CLASSIFIER",
          "LLM",
          "EMBEDDER",
          "HUBSPOT"
        ],
        "title": "Dependency",
        "type": "string"
      },
      "DiscoveryCandidate": {
        "additionalProperties": false,
        "description": "[`DiscoveryCandidate`](/architecture/interfaces.md#discoverycandidate).",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "employee_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Count"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscoveryCandidateEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "fit_estimate": {
            "title": "Fit Estimate",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/DiscoveryCandidateOrigin"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          },
          "service_id": {
            "title": "Service Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DiscoveryCandidateStatus"
          }
        },
        "required": [
          "id",
          "service_id",
          "name",
          "domain",
          "country_code",
          "industry",
          "employee_count",
          "origin",
          "status",
          "fit_estimate",
          "evidence",
          "reject_reason",
          "account_id"
        ],
        "title": "DiscoveryCandidate",
        "type": "object"
      },
      "DiscoveryCandidateEvidence": {
        "additionalProperties": false,
        "description": "`DiscoveryCandidate.evidence`.",
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "url",
          "published_at",
          "quote"
        ],
        "title": "DiscoveryCandidateEvidence",
        "type": "object"
      },
      "DiscoveryCandidateOrigin": {
        "description": "`discovery_candidate.origin`.",
        "enum": [
          "CRUNCHBASE_SEARCH",
          "NEWS_MENTION"
        ],
        "title": "DiscoveryCandidateOrigin",
        "type": "string"
      },
      "DiscoveryCandidateStatus": {
        "description": "`discovery_candidate.status`.",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "REJECTED"
        ],
        "title": "DiscoveryCandidateStatus",
        "type": "string"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "`error` of [Conventions](/architecture/interfaces.md#conventions) `ErrorEnvelope`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails",
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetailField": {
        "additionalProperties": false,
        "description": "One entry of `details.fields[]` ([Conventions](/architecture/interfaces.md#conventions)\n`VALIDATION`).",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetailField",
        "type": "object"
      },
      "ErrorDetails": {
        "additionalProperties": false,
        "description": "`details` of [Conventions](/architecture/interfaces.md#conventions) `ErrorEnvelope`: the\nclosed set of names the Envelope table gives. Every field is optional and, per the Naming\nparagraph (G10), absent rather than `null` when unset.",
        "properties": {
          "dependency": {
            "$ref": "#/components/schemas/Dependency"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetailField"
            },
            "title": "Fields",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resets_at": {
            "format": "date-time",
            "title": "Resets At",
            "type": "string"
          },
          "retry_after_min": {
            "title": "Retry After Min",
            "type": "integer"
          }
        },
        "title": "ErrorDetails",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The `{\"error\": {...}}` shape of [Conventions](/architecture/interfaces.md#conventions),\nnamed `ErrorEnvelope` in the OpenAPI document.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Page_DiscoveryCandidate_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DiscoveryCandidate"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[DiscoveryCandidate]",
        "type": "object"
      },
...
      "PipelineRunKind": {
        "description": "`pipeline_run.kind`.",
        "enum": [
          "ACCOUNT_REFRESH",
          "RECLASSIFY",
          "RESCORE",
          "DISCOVERY",
          "EVALUATION"
        ],
        "title": "PipelineRunKind",
        "type": "string"
      },
      "PipelineRunStage": {
        "description": "`pipeline_run.stage`.",
        "enum": [
          "FETCH",
          "PROCESS",
          "TRIAGE",
          "CLASSIFY",
          "EVIDENCE",
          "SCORE"
        ],
        "title": "PipelineRunStage",
        "type": "string"
      },
      "PipelineRunStatus": {
        "description": "`pipeline_run.status`.",
        "enum": [
          "QUEUED",
          "RUNNING",
          "SUCCEEDED",
          "PARTIAL",
          "FAILED",
          "CANCELLED"
        ],
        "title": "PipelineRunStatus",
        "type": "string"
      },
      "PipelineRunTrigger": {
        "description": "`pipeline_run.trigger`.",
        "enum": [
          "SCHEDULE",
          "USER",
          "QUESTION_CHANGE",
          "SCORING_ACTIVATION",
          "ACCOUNT_CHANGE",
          "FEEDBACK",
          "OVERRIDE"
        ],
        "title": "PipelineRunTrigger",
        "type": "string"
      },
...
      "Run": {
        "additionalProperties": false,
        "description": "[`Run`](/architecture/interfaces.md#run).",
        "properties": {
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "ai_cost_eur": {
            "title": "Ai Cost Eur",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RunError"
            },
            "title": "Errors",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PipelineRunKind"
          },
          "progress": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Progress",
            "type": "object"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunQuestion"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Name"
          },
          "service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineRunStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineRunStatus"
          },
          "trigger": {
            "$ref": "#/components/schemas/PipelineRunTrigger"
          }
        },
        "required": [
          "id",
          "kind",
          "trigger",
          "status",
          "stage",
          "progress",
          "errors",
          "account",
          "service",
          "question",
          "requested_by_name",
          "created_at",
          "started_at",
          "finished_at",
          "ai_cost_eur"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunError": {
        "additionalProperties": false,
        "description": "One entry of `Run.errors`, as [`pipeline_run`](/architecture/sql-store.md#pipeline_run)\n`errors` states it.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "plugin_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourcePluginCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "$ref": "#/components/schemas/PipelineRunStage"
          }
        },
        "required": [
          "stage",
          "code",
          "message"
        ],
        "title": "RunError",
        "type": "object"
      },
      "RunQuestion": {
        "additionalProperties": false,
        "description": "`Run.question`: `{id, key}`.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key"
        ],
        "title": "RunQuestion",
        "type": "object"
      },
      "RunRef": {
        "additionalProperties": false,
        "description": "`Run.account` and `Run.service`: `{id, name}`.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "RunRef",
        "type": "object"
      },
...
      "SourcePluginCode": {
        "description": "`source_plugin.code`; reused by `plugin_usage.plugin_code` and `document.plugin_code`.",
        "enum": [
          "GDELT",
          "RSS",
          "WEBSITE",
          "CAREERS",
          "CRUNCHBASE",
          "NEWSAPI",
          "SERPAPI"
        ],
        "title": "SourcePluginCode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "LeadRadar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/discovery-candidates": {
      "get": {
        "description": "`API-30`.",
        "operationId": "list_discovery_candidates_api_v1_discovery_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DiscoveryCandidateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DiscoveryCandidate_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List Discovery Candidates",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/v1/discovery-candidates/{id}/accept": {
      "post": {
        "description": "`API-31`.",
        "operationId": "accept_discovery_candidate_api_v1_discovery_candidates__id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Accept Discovery Candidate",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/v1/discovery-candidates/{id}/reject": {
      "post": {
        "description": "`API-32`.",
        "operationId": "reject_discovery_candidate_api_v1_discovery_candidates__id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryCandidate"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Reject Discovery Candidate",
        "tags": [
          "discovery"
        ]
      }
    },
...
    "/api/v1/services/{id}/discovery-runs": {
      "post": {
        "description": "`API-29`.",
        "operationId": "start_discovery_run_api_v1_services__id__discovery_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Start Discovery Run",
        "tags": [
          "discovery"
        ]
      }
    },
...
  }
}